feat(metagame)!: widen mint metadata to u128, matching the batch path
`mint` took `metadata: u16` while `mint_batch_recipients` took `u128`, so the
two paths disagreed about how much metadata a caller could carry. Budokan
threads a u128 metadata_value through all three of its entrypoints; adopting
`mint` as-is would have narrowed its single-mint path to 16 bits while its
batch path kept 128 — a regression on work that had just landed.

BREAKING: `libs::mint`, `MetagameComponent::mint` and
`MintMetagameParams.metadata` all take `u128`. The legacy token's field is
still `u16`, so the legacy branch narrows with the same
`try_into().expect('Metagame: metadata exceeds u16')` the batch path already
used — both paths now reject identically rather than one truncating.

Only the metagame surfaces widened. Most `metadata: u16` occurrences in these
files belong to legacy TOKEN mocks implementing `IMinigameTokenLegacy`, whose
field genuinely is u16; widening those would have made the mocks lie about the
interface they implement. `IMinigame::mint_game` also keeps u16 — the minigame
helper layer is legacy-only (MinigameComponent asserts
IMINIGAME_TOKEN_LEGACY_ID), so u16 is correct there.

Two tests that discriminate rather than decorate: `test_mint_carries_wide_metadata`
round-trips 0x100000000 through `mint` and reads it back via `mint_metadata`
(would not compile under u16, would fail under silent truncation), and
`test_mint_rejects_wide_metadata_on_legacy_token` proves the single path
narrows exactly as the batch path does.

Raised by the budokan session, which hit it while evaluating adoption.

Full package: 1212 passed / 0 failed, no compilation diagnostics, fmt clean.

+++ packages/embeddable_game_standard/src/metagame/metagame_component.cairo
@@ -48,7 +48,7 @@ pub mod MetagameComponent {
             soulbound: bool,
             paymaster: bool,
             salt: u16,
-            metadata: u16,
+            metadata: u128,
         ) -> felt252 {
             libs::mint(
                 game_address,
