+++ packages/embeddable_game_standard/src/metagame/tests/test_tournament_flow.cairo
@@ -273,7 +273,7 @@ mod MockMetagameWithContext {
             soulbound: bool,
             paymaster: bool,
             salt: u16,
-            metadata: u16,
+            metadata: u128,
         ) -> felt252 {
             self
                 .metagame
@@ -317,7 +317,7 @@ trait IMockMetagame<TContractState> {
         soulbound: bool,
         paymaster: bool,
         salt: u16,
-        metadata: u16,
+        metadata: u128,
     ) -> felt252;
 }
